fix: Remove narrowing conversion of types

+++ solution/include/bmp_io.h
@@ -13,7 +13,7 @@ enum read_status from_bmp(FILE *in, struct image *img);
 
 enum write_status to_bmp(FILE *out, const struct image *img);
 
-size_t bmp_image_get_padding_in_bytes(const struct image *img);
+int64_t bmp_image_get_padding_in_bytes(const struct image *img);
 
 static const char *const image_read_error_msg[] = {
     [READ_OK] = "File read successfully!",
